Adding a context summarizer to the execution agent
In order to avoid running into problems with the context window in long execution agent runs, Ive added a context summarizer usig the LangChain SummarizerMiddleware.

When the (approximate) number of tokens in the message history > tokens_before_summarize (default 50000), the summarizer will take all messages older than "messages_to_keep" (default 20) and use the LLM to summarize them, keeping as much information as possible.

I tested this with the values set to 20000 and 10 for the fibonacci plan-execute problem and it worked well, reducing the context length from > 20000 to ~4000 each time it summarized. This resulted in a run that never threatened to run out of tokens and cost half as much as previously

It could probably use more testing, but this should address Nathans problems.