Got a version of an add_mcp_tools method working.
It required integrating asyncio into URSA
because the MCP tool calling in langchain requires it. So there is an _ainvoke() method
defined in the ExecutionAgent (and in the BaseAgent) and a .ainvoke in the BaseAgent.

This did not seem to break any functionality without using asyncio, but does require
using it for the MCP calls. Will not integrate with the CLI smoothly yet I dont think,
but we can talk about that later.