Add tools method (#96)
* Part way through implementing some tool adding functionality to the execution agent.

* Got a version of an add_mcp_tools method working.

It required integrating asyncio into URSA
because the MCP tool calling in langchain requires it. So there is an _ainvoke() method
defined in the ExecutionAgent (and in the BaseAgent) and a .ainvoke in the BaseAgent.

This did not seem to break any functionality without using asyncio, but does require
using it for the MCP calls. Will not integrate with the CLI smoothly yet I dont think,
but we can talk about that later.

* Removing an extraneous print statement.

* Removing two extraneous prints and removing abstractmethod from _ainvoke()

* update uv.lock

* Small bug fix for this.

* replace _INVOKE_DEPTH with BaseAgent._invoke_depth

* ignore ursa_metrics

* deduplicate code in invoke and ainvoke

* add extra_tools and docs/tests for it

---------

Co-authored-by: Arthur Lui <alui@lanl.gov>