Adding a context summarizer to the execution agent (#95)
* Adding a context summarizer to the execution agent

Add a context summarizer using the LangChain SummarizerMiddleware
to avoid exceeding context window in long execution agent runs.

When the (approximate) number of tokens in the message history > tokens_before_summarize (default 50000), the summarizer will take all messages older than "messages_to_keep" (default 20) and use the LLM to summarize them, keeping as much information as possible.

I tested this with the values set to 20000 and 10 for the fibonacci plan-execute problem and it worked well, reducing the context length from > 20000 to ~4000 each time it summarized. This resulted in a run that never threatened to run out of tokens and cost half as much as previously

It could probably use more testing, but this should address Nathans problems.

* Should be saving out new_state, not something based on the original state.

* Removing extra print statements and replaced with a rich console print that happens whenever the summarizing happens.

* explicit summarize middleware params

* dry summarize context

* Fixing a bug where no new message was appended when there was an LLM Response issue and so the system message could be the most recent message to the should_continue check. Appending an AIMessage with the content of the error. This ensures that the most recent message can be checked for tool_calls.

* check that system message is first

---------

Co-authored-by: Arthur Lui <alui@lanl.gov>