Removing some defaults to avoid unnecessary OpenAI dependency
Some default models and embeddings led to the OPENAI_API_KEY variable
needing to be set even if it was not being used. I removed those defaults
so that users using local models or non OpenAI models do not need to set this.

Previously that env could be set to an empty string when not using OpenAI
but that was sloppy.